pyproject.toml
[project]
name = "pydantic-ai-guardrails"
dynamic = ["version"]
description = "Production-ready guardrails for Pydantic AI with native integration patterns"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = [
    "pydantic-ai",
    "guardrails",
    "ai-safety",
    "llm",
    "validation",
    "input-validation",
    "output-validation",
    "pii-detection",
    "prompt-injection",
    "ai-security",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic-ai>=0.0.1",
    "typing-extensions>=4.12.0",
    "anyio>=4.0.0",
]

[[project.authors]]
name = "Jag Reehal"
email = "jag@jagreehal.com"

[project.urls]
Homepage = "https://github.com/jagreehal/pydantic-ai-guardrails"
Documentation = "https://github.com/jagreehal/pydantic-ai-guardrails#readme"
Repository = "https://github.com/jagreehal/pydantic-ai-guardrails"
Issues = "https://github.com/jagreehal/pydantic-ai-guardrails/issues"

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.0",
    "mypy>=1.18.0",
]
# Optional dependencies for specific guardrails
pii-detection = [
    "presidio-analyzer>=2.2.0",
    "presidio-anonymizer>=2.2.0",
]
toxicity-detection = [
    "detoxify>=0.5.0",
]
prompt-injection = [
    "transformers>=4.30.0",
]
telemetry = [
    "opentelemetry-api>=1.38.0",
    "opentelemetry-sdk>=1.38.0",
]
all = [
    "pydantic-ai-guardrails[pii-detection,toxicity-detection,prompt-injection,telemetry]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pydantic_ai_guardrails"]

[tool.hatch.version]
path = "src/pydantic_ai_guardrails/__version__.py"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/testing/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]

[tool.ruff]
line-length = 120
target-version = "py310"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "SIM102", # nested if statements (sometimes clearer than combined)
    "SIM105", # contextlib.suppress (try-except-pass is clearer in examples)
    "SIM108", # ternary operators (if-else can be clearer)
    "ARG001", # unused arguments (required by signatures, especially in examples)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = false  # More lenient for now
disallow_subclassing_any = true
disallow_untyped_calls = false  # More lenient for now
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
exclude = [
    "examples/",  # Exclude example files that use third-party libraries without type stubs
    "venv/",      # Exclude virtual environment
]

# Relax type checking for tests
[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false
disable_error_code = ["arg-type", "assignment", "union-attr", "operator", "var-annotated", "dict-item"]
